change project name and code struct
