+++ pyproject.toml
@@ -1,10 +1,10 @@
 [tool.poetry]
-name = "dyn-cli"
+name = "dyna-cli"
 version = "0.1.0"
 description = ""
 authors = ["Rowan Self <piesrule123@gmail.com>"]
 readme = "README.md"
-packages = [{include = "dyn_cli"}]
+packages = [{include = "dyna_cli"}]
 
 [tool.poetry.dependencies]
 python = "^3.9"
@@ -12,7 +12,6 @@ textual = "^0.15.1"
 boto3 = "^1.26.96"
 simplejson = "^3.18.4"
 dynamodb-json = "^1.3"
-pandas = "^1.5.3"
 
 
 [tool.poetry.group.dev.dependencies]
